+++ stroom-config/stroom-config-global-impl/src/main/java/stroom/config/global/impl/ConfigMapper.java
@@ -35,6 +35,7 @@ import stroom.util.config.annotations.Password;
 import stroom.util.config.annotations.ReadOnly;
 import stroom.util.config.annotations.RequiresRestart;
 import stroom.util.logging.LogUtil;
+import stroom.util.shared.ConfigValidationResults;
 import stroom.util.shared.IsConfig;
 
 import javax.inject.Inject;
@@ -93,6 +94,10 @@ public class ConfigMapper {
     // A map of property accessor objects keyed on the fully qualified prop path (i.e. stroom.path.temp)
     private final Map<String, Prop> propertyMap = new HashMap<>();
 
+    // A map of the config object instance to its property path, e.g. <nodeCondfig> => 'stroom.node'
+    // Allows us to find the path of a config object
+    private final Map<IsConfig, String> configInstanceToPathMap = new HashMap<>();
+
     @Inject
     public ConfigMapper(final AppConfig appConfig) {
         LOGGER.debug("Initialising ConfigMapper with class {}", appConfig.getClass().getName());
@@ -114,6 +119,7 @@ public class ConfigMapper {
                 vanillaObject,
                 ROOT_PROPERTY_PATH,
                 new HashMap<>(),
+                new HashMap<>(),
                 this::defaultValuePropertyConsumer);
 
         } catch (InstantiationException
@@ -133,7 +139,7 @@ public class ConfigMapper {
      * and update the globalPropertiesMap.
      * It will also apply common database config to all other database config objects.
      */
-    void updateConfigFromYaml(final AppConfig newAppConfig) {
+    private void updateConfigFromYaml(final AppConfig newAppConfig) {
         synchronized (this) {
             if (System.identityHashCode(newAppConfig) != System.identityHashCode(appConfig)) {
                 // We have been passed a different object to our instance appConfig so copy all
@@ -145,12 +151,29 @@ public class ConfigMapper {
             // Now walk the AppConfig object model from the DropWiz YAML updating globalPropertiesMap with the
             // YAML values where present.
             LOGGER.debug("Adding yaml config values into global property map");
-            addConfigObjectMethods(newAppConfig, ROOT_PROPERTY_PATH, propertyMap, this::yamlPropertyConsumer);
+            addConfigObjectMethods(
+                newAppConfig, ROOT_PROPERTY_PATH, propertyMap, configInstanceToPathMap, this::yamlPropertyConsumer);
         }
     }
 
     public boolean validatePropertyPath(final String fullPath) {
         return propertyMap.get(fullPath) != null;
+    }
+
+    public ConfigValidationResults validateConfiguration() {
+        final ConfigValidationResults.Aggregator aggregator = ConfigValidationResults.aggregator();
+        for (final Map.Entry<IsConfig, String> entry : configInstanceToPathMap.entrySet()) {
+            aggregator.addAll(entry.getKey().validateConfig());
+        }
+        return aggregator.aggregate();
+    }
+
+    public String getBasePath(final IsConfig config) {
+        final String path = configInstanceToPathMap.get(config);
+
+        Objects.requireNonNull(path, LogUtil.message("Class {} has no property path",
+            config.getClass().getCanonicalName()));
+        return path;
     }
 
     Collection<ConfigProperty> getGlobalProperties() {
@@ -217,8 +240,12 @@ public class ConfigMapper {
     private void addConfigObjectMethods(final IsConfig object,
                                         final String path,
                                         final Map<String, Prop> propertyMap,
+                                        final Map<IsConfig, String> configInstanceToPathMap,
                                         final BiConsumer<String, Prop> propConsumer) {
         LOGGER.trace("addConfigObjectMethods({}, {}, .....)", object, path);
+
+        // Add our object with its path to the map
+        configInstanceToPathMap.put(object, path);
 
         final Map<String, Prop> properties = PropertyUtil.getProperties(object);
         properties.forEach((k, prop) -> {
@@ -245,7 +272,8 @@ public class ConfigMapper {
                 // This must be a branch, i.e. config object so recurse into that
                 if (value != null) {
                     IsConfig childConfigObject = (IsConfig) value;
-                    addConfigObjectMethods(childConfigObject, fullPath, propertyMap, propConsumer);
+                    addConfigObjectMethods(
+                        childConfigObject, fullPath, propertyMap, configInstanceToPathMap, propConsumer);
                 }
             } else {
                 // This is not expected