+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Add validation of the config file on boot.
+
 * Change docker image to allow us to pass in the dropwizard command to run, e.g. server|migrate.
 
 * Stop MySQL outputing Note level warnings during migration about things that don't exist when we expect them not to.